vertico.el: hack C-k to remove entries from recentf-list
* src/vertico.el: (vertico-kill-buffer): special case
consult-recent-files so C-k removes entries from the
recentf-list. This does not work as well as C-k to kill buffers
because of limitations in the design of consult: as it prepares the
list of candidates upfront, and cannot adapt to changes, so vertico
ends up completing on the older list. However The next time you call
consult-recentf-files it picks up the new recentf-list.

The other problem is consult-recent-files calls abbreviate-file-name
on the set of candidates, so what is being shown by vertico need not
be the same entry in the recentf-list. The sub-optimal workaround is
to delete the both expanded file name as well as the abbreviated file
name. This may not be what the user always wants (i.e. he wants to
delete one but not the other)

TODO: rename vertico-kill-buffer to reflect the added
responsibilities.

[240317] extend to include find-file family (for file-name-history)
and consult-buffer. unfortunately consult-dir won't work.

[240726] extend to include support for consult-recent-file: remove
entries from dired-recent-directories (via dired-recent.el and
consult.el) -- consult-recent-file is assumed to have been modified to
append dired-recent-directories to recentf-files.