vertico.el: (vertico--recompute): revert BREAKING CHANGE: Always preselect the first candidate
;madhu 030208
The BREAKING CHANGE was reverted in 452f1c931f090498
but we still have to support always-preselect-first-candidate

;madhu 021219
* vertico.el: (vertico-always-preselect-first-candidate): new
variable. default to NIL which was the earlier
behaviour. (vertico--recompute): use it here.

This commit partially undoes the breaking change in 565faea4b7686.  To
get standard vertico behaviour set
vertico-always-preselect-first-candidate to T.

The old behaviour is especially needed when calling find-file with
ffap-supersede and one can toggle between dired and find-file.

;; change dired-at-prompt-prompter to prompt the guessed filename. not
;; just the directory

;madhu 240603
does 'never-prompt make this patch unnecessary?
no time to check now.