vertico.el: (vertico-setup): don't ignore additional minibuffer keybinings
vertico-map inherits minibuffer-local-map but minibuffer completion
sets up the local keymap with more than just the minibuffer-local-map.
to wit it uses minibuffer-local-completion-map,
minibbuffer-local-filename-completion-map, and
minibuffer-local-must-match-map to set up the local keymap.

this patch should preserve the additional keybindings that minibuffer
completion sets up so they are available when using vertico.

[230911]
Unfortunately vertico has chosen to use the new defvar-keymap
interface which remaps bindings in minibuffer-local-map.

To support the behaviour proposed in this patch, it would be better
define vertico-map without a parent keymap so it could composed with
the current-local-map which is set up by the minibuffer-completion
mechanism (which includes minibuffer-local-map)